+++ src/Requests/Transactions/ConfirmTransactionRequest.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Mobian\ApiClient\Requests\Transactions;
+
+use Mobian\ApiClient\Requests\AbstractRequest;
+
+class ConfirmTransactionRequest extends AbstractRequest
+{
+    /**
+     * @inheritdoc
+     */
+    protected $method = self::METHOD_POST;
+
+    /**
+     * @inheritdoc
+     */
+    protected $endpoint = '/api/transactions/%d/confirm';
+
+    /**
+     * Constructor.
+     *
+     * @param int $identifier
+     */
+    public function __construct(int $identifier)
+    {
+        $this->endpoint = sprintf($this->endpoint, $identifier);
+    }
+}