+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] - 2020-08-10
+
+**Added**
+
+* Discount code check request.
+* Confirm transaction request.
+
+
 ## [1.0.3] - 2020-05-25
 
 **Added**
@@ -140,6 +148,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * Amend booking request.
 * Catalogue requests for taxi and transfer products.
+
 
 ## [0.3.0] - 2018-11-16